add a login service to clean up controller
