+++ aspnet-api/PortfolioBack/Controllers/LoginController.cs
@@ -3,77 +3,41 @@ using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Security.Claims;
-using PortfolioBack.Data;
 using PortfolioBack.DTOs;
-using PortfolioBack.Models;
 using PortfolioBack.Services;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.IdentityModel.Tokens;
-using System.Text;
-using System.IdentityModel.Tokens.Jwt;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
 
 [ApiController]
 [Route("api/[controller]")]
-public class LoginController(PortfolioDbContext db, IPasswordHasher hasher, IConfiguration configuration) : ControllerBase
+public class LoginController(LoginService loginService) : ControllerBase
 {
-
   // [HttpPost("signup")]
   // [AllowAnonymous]
   // public async Task<ActionResult<AuthUserDto>> Signup([FromBody] SignupRequestDto request)
   // {
-  //   var username = request.Username.Trim();
-  //   var exists = await db.Users.AnyAsync(u => u.Username == username);
-  //   if (exists) return Conflict(new { message = "Username already exists" });
-
-  //   var (hash, salt, iterations) = hasher.HashPassword(request.Password);
-  //   var user = new User
-  //   {
-  //     Username = username,
-  //     PasswordHash = hash,
-  //     PasswordSalt = salt,
-  //     PasswordIterations = iterations
-  //   };
-  //   db.Users.Add(user);
-  //   await db.SaveChangesAsync();
-
-  //   var token = CreateToken(user);
-  //   return Ok(new AuthUserDto {
-  //     Id = user.Id,
-  //     Username = user.Username,
-  //     Token = token
-  //   });
+  //   var authUserDto = loginService.Signup(request);
+  //   if (authUserDto is null) return Conflict(new { message = "Username already exists" });
+  //   return Ok(authUserDto);
   // }
 
   [HttpPost("login")]
   [AllowAnonymous]
   public async Task<ActionResult<AuthUserDto>> Login([FromBody] LoginRequestDto request)
   {
-    var username = request.Username.Trim();
-    var user = await db.Users.SingleOrDefaultAsync(u => u.Username == username);
-    if (user == null)
+    var authUserDto = await loginService.Login(request);
+    if (authUserDto is null)
     {
       await Task.Delay(Random.Shared.Next(50, 150)); // timing noise
       return Unauthorized(new { message = "Invalid credentials" });
     }
-
-    var valid = hasher.Verify(request.Password, user.PasswordHash, user.PasswordSalt, user.PasswordIterations);
-    if (!valid) return Unauthorized(new { message = "Invalid credentials" });
-
-    var token = CreateToken(user, configuration);
-    return Ok(new AuthUserDto
-    {
-      Id = user.Id,
-      Username = user.Username,
-      Token = token
-    });
+    return Ok(authUserDto);
   }
 
   [HttpPost("logout")]
   [Authorize]
-  [ValidateAntiForgeryToken]
   public async Task<IActionResult> Logout()
   {
-    await HttpContext.SignOutAsync(CookieAuthenticationDefaults.AuthenticationScheme);
+    await HttpContext.SignOutAsync(JwtBearerDefaults.AuthenticationScheme);
     return NoContent();
   }
 
@@ -91,41 +55,5 @@ public class LoginController(PortfolioDbContext db, IPasswordHasher hasher, ICon
     });
   }
 
-  private string CreateToken(User user, IConfiguration configuration)
-  {
-    var claims = new List<Claim>
-    {
-      new Claim(ClaimTypes.Name, user.Username),
-      new Claim(ClaimTypes.NameIdentifier, user.Id.ToString())
-    };
 
-    var jwtKey = configuration.GetValue<string>("Jwt:Key");
-    var jwtIssuer = configuration.GetValue<string>("Jwt:Issuer");
-    var jwtAudience = configuration.GetValue<string>("Jwt:Audience");
-    var jwtExpMinutes = configuration.GetValue<string>("Jwt:AccessTokenExpirationMinutes");
-
-    if (new List<string?>
-    {
-        jwtAudience, jwtIssuer, jwtKey, jwtExpMinutes
-    }.Any(s => s == null))
-    { throw new NullReferenceException("Jwt settings are missing"); }
-
-    var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(jwtKey!));
-
-    var credentials = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
-
-    int minutes;
-    if (!Int32.TryParse(jwtExpMinutes, out minutes))
-    {
-      minutes = 30;
-    }
-    var tokenDescriptor = new JwtSecurityToken(
-      issuer: jwtIssuer,
-      audience: jwtAudience,
-      claims: claims,
-      expires: DateTime.UtcNow.AddMinutes(minutes),
-      signingCredentials: credentials);
-
-    return new JwtSecurityTokenHandler().WriteToken(tokenDescriptor);
-  }
 }
